problem with loading css partial routes fixed
