+++ routes/index.js
@@ -1,9 +1,16 @@
 var express = require('express');
 var router = express.Router();
 
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express' });
-});
+router.get('/', (req, res) => {
+  res.render('index.ejs');
+})
+
+router.get('/login', (req, res) => {
+  res.render('login.ejs');
+})
+
+router.get('/register', (req, res) => {
+  res.render('register.ejs');
+})
 
 module.exports = router;
