+++ views/index.ejs
@@ -1,32 +1,24 @@
-<!DOCTYPE html>
-<html lang="en">
+<%- include ('partials/head.ejs') %>
 
-<head>
-    <%- include ('partials/head.ejs') %>
-</head>
 
 <body class="container">
+  <%- include ('partials/header.ejs') %>
 
-    <header>
-        <%- include ('partials/header.ejs') %>
-    </header>
+  <main>
 
-    <main>
+    <section id="slider">
 
-        <section id="slider">
+      <h1> Swipe. Match.
+        <span href="" class="typewrite" data-period="2000" data-type='[ " Enjoy", " Love", " Date", " Chat" ]'></span>
+      </h1>
 
-            <h1> Swipe. Match.
-                <span href="" class="typewrite" data-period="2000"
-                    data-type='[ " Enjoy", " Love", " Date", " Chat" ]'></span>
-            </h1>
+      <p>Waarom ingewikkeld? Maak daten lekker makkelijk met Lovr!</p>
 
-            <p>Waarom ingewikkeld? Maak daten lekker makkelijk met Lovr!</p>
+      <button onclick="location.href='register'" type="button">Aanmelden</button>
 
-            <button onclick="location.href='aanmelden'" type="button">Aanmelden</button>
+    </section>
 
-        </section>
-
-    </main>
+  </main>
 
 </body>
