+++ views/login.ejs
@@ -33,7 +33,7 @@
 
         <div class="box-secondary">
           <p>Nog geen account?</p>
-          <button><a href="/registreren">Registreren</a></button>
+          <button><a href="/register">Registreren</a></button>
         </div>
 
       </div>
