+++ views/partials/head.ejs
@@ -1,20 +1,21 @@
 <!-- views/partials/head.ejs -->
-<meta charset="UTF-8">
-<meta name="description" content="Mobile Dating Applicatie voor CMDBT-19-20">
-<meta name="keywords" content="Lovr, Dating, Students, CMD, HvA">
-<meta name="author" content="Azam Awan - Klas: 203">
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
 
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
 
-<!-- Externe CSS pagina inladen -->
-<link rel="stylesheet" type="text/css" href="/public/stylesheets/style.css">
+  <!-- Externe CSS pagina inladen -->
+  <link rel="stylesheet" type="text/css" href="./public/stylesheets/style.css">
 
-<!-- Externe JQuery pagina inladen -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+  <!-- Externe JQuery pagina inladen -->
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
 
-<!-- app.js inladen-->
-<script src="/public/js/script.js"></script>
+  <!-- app.js inladen-->
+  <script src="./public/js/script.js"></script>
 
-<!-- Font Awesome 4 iconografie inladen -->
-<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css"
-     integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
+  <!-- Font Awesome 4 iconografie inladen -->
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css"
+       integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
+</head>
